// src/types/enums/MessageUnum.ts
export enum ERROR {
    SERVER_ERROR = "There is error in server!",
    USER_EXIST = "This user already exists!",
    EMAIL_EXIST = "This email already exists!",
    CANT_FIND_USER = "This user cannot be found!",
    CANT_FIND_HANDLER = "This admin cannot be found!",
    CANT_FIND_PROPERTY = "This property cannot be found!",
    CANT_FIND_BLog = "This blog cannot be found!",
    CANT_FIND_TESTIMONIALS = "This Testimonials cannot be found!",
    INVALID_DATA = "please enter a valid data!",
    INVALID_EMAIL = "please enter a valid email!",
    WRONG_PASSWORD = "Email or password is incorrect!",
    PROBLEM = "There is a Problem!",
    NO_USER = "There is no user!",
    ACCESS_DENIED = "access denied!",
    LOGIN = "Please login first!",
    REQUIRED_FIELD = "This field must be filled.",
    REQUIRED_NAME = "Name is required for sign-up.",
    REQUIRED_LINCENSE_NUMBER = "lincense number is required for create agent.",
    INVALID_EMAIL_FORMAT = "Invalid email format.",
    NAME_ATLEAST = "Name must be at least 3 characters.",
    LASTNAME_ATLEAST = "Last name must be at least 3 characters.",
    PASSWORD_ATLEAST = "Password must be at least 6 characters.",
    PHONE_ATLEAST = "Phone number must be 11 number.",
    PASSWORD_DONT_MACH = "Passwords do not match.",
    RESET_LINK_EXPIRED = "Your reset link has expired. Please request a new one.",
    INVALID_TOKEN = "Invalid token. Please try again.",
    OTP_REQUIRED = "OTP Code is required!",
    UNAUTHORIZED = "You dont have permission to edit this user.",
    ALREADY_PROMOTED = "This user alredy promoted to this level",
    NO_AGENT = "There is error in finding agent!",
    NO_MESSAGE = "There is error in finding messages!",
    NO_FORM = "There is error in finding form messages!",
    CANT_FIND_FAQ = "This question cannot be found!",
    INVALID_MESSAGE = "This message is not related to you!",
    INVALID_FORMAT = "Invalid file format" 
}

export enum MESSAGE {
    NEW_USER = "User cteated successfully!",
    NEW_BLOG_TESTIMONIALS = "Testimonial added successfully!",
    NEW_BLOG_TESTIMONIALS_REPLY = "Testimonial reply added successfully!",
    USER_DELETE = "User deleted successfully!",
    PASSWORD_RESET_EMAIL = "Password reset email sent successfully.",
    PASSWORD_RESET = "Password reset successfully.",
    VALID_TOKEN = "Valid token. you can reset your password.",
    PROFILE_EDIT = "Your profile edited successfully.",
    NEW_ADMIN = 'new Admin promoted successfully',
    NEW_AGENT = 'new Admin promoted successfully',
    NEW_MESSAGE = "New message have been sent successfully ",
    NEW_PROPERTY = "New property added successfully ",
    NEW_BLOG = "Blog created successfully",
    PROPERTY_EDITED = "property Edited successfully ",
    PROPERTY_DELETED = 'property deleted successfully',
    BLOG_DELETED = 'blog deleted successfully',
    PROPERTY_PUBLISHED = "property Published successfully ",
    BLOG_PUBLISHED = "blog Published successfully ",
    PROPERTY_REJECTED = "property Rejected successfully ",
    Blog_REJECTED = "BLog Rejected successfully ",
    NEW_FORM = "New message have been sent successfully ",
    READ_MESSAGE = " message have been read successfully ",
    READ_FORM = "form message have been read successfully ",
    READ_ALL_MESSAGE = "All message have been read successfully ",
    READ_ALL_Form = "All from message have been read successfully ",
    FAQ_DELETE = "Question deleted successfully!",
    FAQ_EDIT = "Question edtited successfully!",
    NEW_FAQ = "New F&Q added successfully!",
}